+++ kerb4j-server/kerb4j-server-spring-security/src/main/java/com/kerb4j/client/spring/KerberosRestTemplate.java
@@ -1,18 +1,3 @@
-/*
- * Copyright 2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.kerb4j.client.spring;
 
 import com.kerb4j.common.util.Constants;
@@ -26,8 +11,8 @@ import org.springframework.web.client.RestTemplate;
 
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
 import java.util.Base64;
+import java.util.List;
 
 /**
  * {@code RestTemplate} that is able to make kerberos authenticated REST