+++ kerb4j-server/kerb4j-server-spring-webflux/src/test/java/com/kerb4j/server/spring/webflux/WebFluxSecurityConfig.java
@@ -1,18 +1,3 @@
-/*
- * Copyright 2009-2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.kerb4j.server.spring.webflux;
 
 import com.kerb4j.server.spring.SpnegoAuthenticationProvider;
@@ -39,9 +24,6 @@ import org.springframework.security.web.server.authentication.AuthenticationWebF
  * 
  * <p>This configuration demonstrates how to set up Kerberos/SPNEGO authentication
  * in a Spring WebFlux application using the kerb4j WebFlux components.</p>
- *
- * @author GitHub Copilot
- * @since 1.0
  */
 @Configuration
 @EnableWebFluxSecurity